# backend/docs/swagger/swagger.yaml
basePath: /
definitions:
  character.PracticePointResponse:
    properties:
      amount:
        type: integer
      skill_name:
        type: string
    type: object
...
host: localhost:8180
info:
  contact:
    email: bamort.support@trokan.de
    name: BaMoRT Support
  description: |-
    BaMoRT (MOAM Replacement) - Role-playing Game Character Management System API

    This API provides comprehensive character management, import/export functionality,
    master data management, and PDF generation for tabletop role-playing games.

    ## Authentication
    Most endpoints require authentication via JWT token. Obtain a token by calling POST /login.
    Include the token in the Authorization header: `Authorization: Bearer <token>`

    ## Character Import/Export
    The API supports pluggable adapters for importing/exporting characters from various systems.
    Supported formats include MOAM VTT, Foundry VTT, and more via microservice adapters.
  license:
    name: Custom License
    url: https://github.com/Bardioc26/bamort/blob/main/LICENSE
  title: BaMoRT API
  version: "2.0"
paths:
...
  /api/characters/{id}/practice-points:
    get:
      description: Returns all practice points for a character
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Practice points list
          schema:
            items:
              $ref: '#/definitions/character.PracticePointResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get practice points
      tags:
      - Characters
    put:
      consumes:
      - application/json
      description: Updates the practice points for a character (owner only)
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      - description: Practice points data
        in: body
        name: practice_points
        required: true
        schema:
          items:
            $ref: '#/definitions/character.PracticePointResponse'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Practice points updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied - owner only
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update practice points
      tags:
      - Characters
  /api/characters/{id}/practice-points/add:
    post:
      consumes:
      - application/json
      description: Adds a new practice point to a character (owner only)
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      - description: Practice point data
        in: body
        name: practice_point
        required: true
        schema:
          $ref: '#/definitions/character.PracticePointResponse'
      produces:
      - application/json
      responses:
        "200":
          description: Added practice point
          schema:
            $ref: '#/definitions/character.PracticePointResponse'
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied - owner only
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add practice point
      tags:
      - Characters
  /api/characters/{id}/practice-points/use:
    post:
      consumes:
      - application/json
      description: Marks a practice point as used and applies it to skill improvement
        (owner only)
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      - description: Practice point usage data
        in: body
        name: use_data
        required: true
        schema:
          properties:
            practice_point_id:
              type: integer
            skill_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Practice point used successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied - owner only
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Use practice point
      tags:
      - Characters
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"